consumer/requirements.txt
flask
google-cloud-pubsub
requests
google-cloud-storage
google-cloud-bigquery
fastavro
pandas
pyarrow

